2021-11-20 creando modelo avances
