--- models/avance/resolvers.ts
@@ -1,33 +1,0 @@
-import { ModeloAvance } from "./avance";
-
-const resolversAvance = {
-    Query: {
-        Avances: async (parent, args) => {
-            const avances = await ModeloAvance.find()
-                .populate("proyecto")
-                .populate("creadoPor");
-            return avances;
-        },
-        filtrarAvance: async (parents, args) => {
-            const avanceFiltrado = await ModeloAvance.find({
-                proyecto: args.idProyecto,
-            })
-                .populate("proyecto")
-                .populate("creadoPor");
-            return avanceFiltrado;
-        },
-    },
-    Mutation: {
-        crearAvance: async (parents, args) => {
-            const avanceCreado = ModeloAvance.create({
-                fecha: args.fecha,
-                descripcion: args.descripcion,
-                proyecto: args.proyecto,
-                creadoPor: args.creadoPor,
-            });
-            return avanceCreado;
-        },
-    },
-};
-
-export { resolversAvance };
